Remove dupe ropc b2c tests
